Trivial style tuning on last change.
